+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * Support for inferring the PostgreSQL `hstore` data type (flash-gordon)
 * Added inferring for timestamp types with specified precision (flash-gordon)
 
+### Fixed
+
+* Missing primary key now leads to a more meaningful error (flash-gordon)
+
 ## v1.0.0 2017-01-29
 
 This release is based on rom core 3.0.0 with its improved Schema API, which is extended with SQL-specific features.